Drop nodes or edges

Proxies to dynamically drop *multiple* nodes or edges from an already existing graph.

sg_drop_nodes_p(proxy, data, ids, refresh = TRUE, rate = "once")

sg_drop_edges_p(proxy, data, ids, refresh = TRUE, rate = "once")

Arguments

proxy
An object of class sigmajsProxy as returned by sigmajsProxy.

data
A data.frame of nodes or edges.

ids
Column containing ids to drop from the graph.

refresh
Whether to refresh the graph after node is dropped, required to take effect.

rate
Refresh rate, either once, the graph is refreshed after data.frame of nodes is added or at each iteration (row-wise). Only applies if refresh is set to TRUE.

Note

Have the parameters from your initial graph match that of the node you add, i.e.: if you pass size in your initial chart, make sure you also have it in your proxy.

Examples

# NOT RUN {
demo("add-nodes", package = "sigmajs")
demo("add-edges", package = "sigmajs")
# }